Refactor ApplicationController from rubycritic "B" to "A"
This mostly involved extracting some controller concerns.
