+++ config/initializers/status_code_logging.rb
@@ -3,7 +3,7 @@
 ActiveSupport::Notifications.subscribe('process_action.action_controller') do |*args|
   payload = args.extract_options!
   params = payload[:params]
-  next if DavidRunger::LogSkip.should_skip?(params: params)
+  next if LogSkip.should_skip?(params: params)
 
   # Devise annoyingly sets payload[:status] to 401 in some cases even when an exception has occurred
   status_code = payload[:exception].present? ? 500 : payload[:status]
