+++ lib/log_skip.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module DavidRunger::LogSkip
+module LogSkip
   def self.should_skip?(params: {})
     params['new_relic_ping'].present? && ENV['LOG_NEW_RELIC_PINGS'].blank?
   end
